Move functionality to check and save token to App component
